rgb presented bug found and fixed

+++ proj1/proj1/Raytracer.h
@@ -12,15 +12,15 @@ private:
 	cv::Mat result;
 	int imgHeight;
 	int imgWidth;
-	double maxLightLen = 1000000;
-	double MAX_DEPTH = 15;
+	float maxLightLen = 1000000;
+	float MAX_DEPTH = 15;
 public:
 	RayTracer(Camera* c,Scene* s,int height,int width):camera(c),scene(s),imgHeight(height),imgWidth(width){}
 	cv::Mat getImg() { return result; }
 	void showImg();
 	void writeImg();
 
-	Color trace(Ray* r,double length);
+	Color trace(Ray* r,float length);
 	void run();
 };
