+++ proj1/proj1/scene.h
@@ -4,11 +4,11 @@
 
 struct IntersectPoint {
 	Object* obj;
-	double t;
+	float t;
 
 	IntersectPoint(){}
-	IntersectPoint(Object* o,double _t):obj(o),t(_t){}
-	void set(Object* o, double _t) { obj = o; t = _t; }
+	IntersectPoint(Object* o,float _t):obj(o),t(_t){}
+	void set(Object* o, float _t) { obj = o; t = _t; }
 };
 
 class Scene {
@@ -20,11 +20,11 @@ private:
 public:
 	Scene();
 
-	Meterial* getMeterial(double reflect, double refract,double diffuse, double specular);
+	Meterial* getMeterial(float reflect, float refract,float diffuse, float specular);
 	ColorTexture* getColorTexture(Color c);
 	PicTexture* getPicTexture(std::string path);
-	Sphere* getSphere(Meterial* m, Texture* t, Vector3 o, double r);
-	Plane* getPlane(Meterial* m, Texture* t, Vector3 P, Vector3 n, double D);
+	Sphere* getSphere(Meterial* m, Texture* t, Vector3 o, float r);
+	Plane* getPlane(Meterial* m, Texture* t, Vector3 P, Vector3 n, float D);
 
 	void addObj(Object* obj);
 	int getObjCnt();
